c/snapctl,usersession: drive-by: move xdgopenproxy under usersession
